+++ src/services/reducers/rootReducer.js
@@ -1,0 +1,14 @@
+import { combineReducers } from "redux";
+import { burgerConstructorReducer } from "./burger-constructor";
+import { burgerIngredientsReducer } from "./burger-ingredients";
+import { ingredientReducer } from "./ingredient";
+import { orderReducer } from "./order-details";
+
+export const rootReducer = combineReducers({
+  burgerConstructor: burgerConstructorReducer,
+  burgerIngredients: burgerIngredientsReducer,
+  ingredientDetails: ingredientReducer,
+  order: orderReducer,
+});
+
+export default rootReducer;
